+++ Moveable.cpp
@@ -4,20 +4,20 @@ using namespace std;
                         Getter	Setter
 ------------------------------------------------------------*/
 
-int Moveable::get_x() {
+double Moveable::get_x() {
     return position_x;
 }
-int Moveable::get_y() {
+double Moveable::get_y() {
     return position_y;
 }
 string Moveable::get_dir() {
     return direction;
 }
 
-void Moveable::set_x(int x) {
+void Moveable::set_x(double x) {
     position_x = x;
 }
-void Moveable::set_y(int y) {
+void Moveable::set_y(double y) {
     position_y = y;
 }
 void Moveable::set_dir(string dir) {
